move SmartSink to logfiles, fix a bug where def_path for logs was defined to early and add new function for getting the task-log path
